Fix ONB_OriginateIcmp6Packet_Type2Code0 and add doc comments
o) ComputeTransportChecksum must take as parameters the transport buffer and the ipv6 header of the packet that is being created, not of the failing packet (i.e. this is a "packet too big" icmp6 packet)

o) the destination buffer (icmp6 packet) allocation was of too few bytes: memory corruption could sometimes happen.