Add functionality for reference counting
improvement for locking: when we use an object O over a long period of time,
but only at several points, we should do this:
- use a read lock to retrieve a valid pointer p for object O
- reference the object O so that it would not be deleted after unlock
- unlock

Here, whenever we will need to access the fields a, b and c of object O,
we will use a rw lock of O. However, even when no lock is being held,
we can have the assurance that the pointer p will point to a valid object.

Deletion of a referenced object does not always happen instantaneously:
it happens when all references to it have been removed (i.e. refCount is 0).