On debug mode, we should have OVS_USE_RCU_CALL_STACK and OVS_USE_ASSERTS set
By default, we should have all code checkings enabled in debug mode.

Also, if OVS_USE_RCU_CALL_STACK is set, a "call stack" is registered in the OVS_RCU struct of the objects, so that, if a deference is missing, it would be easier to track the problem.

A function name is saved for the current thread (array of ref calls) in the rcu struct whenever a reference is done, and it is removed from it whenever a dereference happens.