Improve code readability in pers port find funcs
o) Rename pPort -> pOutPort: it shows that it's the port we return
o) Use a local variable instead of pPorts->portsArray[i], when needed in more places
o)Use LIST_FOR_EACH instead of LIST_FOR_EACH_ENTRY: it's shorter
and more straightforward. It assumes the list entry member in struct is listEntry, which is the convention for the structs in this driver.