Use macros for locking the datapath
the macros: DATAPATH_LOCK_READ, DATAPATH_LOCK_WRITE, DATAPATH_UNLOCK
reasons:
o) better clarity when studying the code for locks and unlocks, especially
when there are multiple objects locked in the function, or the same object
is locked and unlocked for multiple times.
o) it might help for debugging, if we want to add custom behavior / debugging
info, whenever a datapath is being locked / unlocked.