Use macros for locking the forward info object
the macros: FWDINFO_LOCK_READ, FWDINFO_LOCK_WRITE, FWDINFO_UNLOCK
reasons:
o) better clarity when studying the code for locks and unlocks, especially
when there are multiple objects locked in the function, or the same object
is locked and unlocked for multiple times.
o) it might help for debugging, if we want to add custom behavior / debugging
info, whenever the forward info is being locked / unlocked.