Fix bug in WinlFlow.c: wrongs flags checked for Flow_Set
Only when both the flags CREATE and EXCLUSIVE are set we are not allowed to overwrite the flow.

Also, small one: if the flow could not be overwritten because it did not match the unmasked key, it is not an error. it simply means a flow was not found for the packet info.