fix for encapsulation: delivery ipv4 identification field
use unique values for delivery ipv4 identification, so that, if the encapsulated buffers get fragmented in their way to the destination, the reassembly will be done correctly.

TODO: this is still not 100% correct, since there is an extremely small possiblity that the system will issue in the same time ipv4 packets with the same identification value, and the same src ip, dst ip, proto type.