Change values for arg types main & invalid
invalid is more natural to be zero, while the specific value of pseudo
group called 'main' is irrelevant (it must only be unique).

This change is needed if we change switch cases to arrays: the default
value of uninitialized values in a static variable is 0, which should
correspond with 'invalid' arg type.

Signed-off-by: Samuel Ghinet <sghinet@cloudbasesolutions.com>