Refactor arg dbgprints for arg types in Argument.c
We've got too many switch cases there, and the code there assumes
that the parent arg type is required in order to correctly dbgprint
an arg type. For a very long time, it is no longer the
case.

We should have only one switch, with all the cases, to find the expected
size for a given arg type.

Signed-off-by: Samuel Ghinet <sghinet@cloudbasesolutions.com>